+++ wit/wit.py
@@ -1,8 +1,9 @@
-import requests
-import os
-import uuid
-import sys
+import json
 import logging
+import os
+import requests
+import sys
+import uuid
 
 WIT_API_HOST = os.getenv('WIT_URL', 'https://api.wit.ai')
 WIT_API_VERSION = os.getenv('WIT_API_VERSION', '20160516')
@@ -16,13 +17,15 @@ class WitError(Exception):
 def req(logger, access_token, meth, path, params, **kwargs):
     full_url = WIT_API_HOST + path
     logger.debug('%s %s %s', meth, full_url, params)
+    headers = {
+        'authorization': 'Bearer ' + access_token,
+        'accept': 'application/vnd.wit.' + WIT_API_VERSION + '+json'
+    }
+    headers.update(kwargs.pop('headers', {}))
     rsp = requests.request(
         meth,
         full_url,
-        headers={
-            'authorization': 'Bearer ' + access_token,
-            'accept': 'application/vnd.wit.' + WIT_API_VERSION + '+json'
-        },
+        headers=headers,
         params=params,
         **kwargs
     )
@@ -53,6 +56,7 @@ def validate_actions(logger, actions):
 class Wit:
     access_token = None
     actions = {}
+    _sessions = {}
 
     def __init__(self, access_token, actions=None, logger=None):
         self.access_token = access_token
@@ -60,16 +64,39 @@ class Wit:
         if actions:
             self.actions = validate_actions(self.logger, actions)
 
-    def message(self, msg, verbose=None):
+    def message(self, msg, context=None, verbose=None):
         params = {}
         if verbose:
             params['verbose'] = True
         if msg:
             params['q'] = msg
+        if context:
+            params['context'] = json.dumps(context)
         resp = req(self.logger, self.access_token, 'GET', '/message', params)
         return resp
 
-    def converse(self, session_id, message, context=None, verbose=None):
+    def speech(self, audio_file, verbose=None, headers=None):
+        """ Sends an audio file to the /speech API.
+        Uses the streaming feature of requests (see `req`), so opening the file
+        in binary mode is strongly reccomended (see
+        http://docs.python-requests.org/en/master/user/advanced/#streaming-uploads).
+        Add Content-Type header as specified here: https://wit.ai/docs/http/20160526#post--speech-link
+
+        :param audio_file: an open handler to an audio file
+        :param verbose:
+        :param headers: an optional dictionary with request headers
+        :return:
+        """
+        params = {}
+        headers = headers or {}
+        if verbose:
+            params['verbose'] = True
+        resp = req(self.logger, self.access_token, 'POST', '/speech', params,
+                   data=audio_file, headers=headers)
+        return resp
+
+    def converse(self, session_id, message, context=None, reset=None,
+                 verbose=None):
         if context is None:
             context = {}
         params = {'session_id': session_id}
@@ -77,15 +104,21 @@ class Wit:
             params['verbose'] = True
         if message:
             params['q'] = message
-        resp = req(self.logger, self.access_token, 'POST', '/converse', params, json=context)
+        if reset:
+            params['reset'] = True
+        resp = req(self.logger, self.access_token, 'POST', '/converse', params,
+                   data=json.dumps(context))
         return resp
 
-    def __run_actions(self, session_id, message, context, i, verbose):
+    def __run_actions(self, session_id, current_request, message, context, i,
+                      verbose):
         if i <= 0:
             raise WitError('Max steps reached, stopping.')
-        json = self.converse(session_id, message, context, verbose)
+        json = self.converse(session_id, message, context, verbose=verbose)
         if 'type' not in json:
             raise WitError('Couldn\'t find type in Wit response')
+        if current_request != self._sessions[session_id]:
+            return context
 
         self.logger.debug('Context: %s', context)
         self.logger.debug('Response type: %s', json['type'])
@@ -124,16 +157,33 @@ class Wit:
                 context = {}
         else:
             raise WitError('unknown type: ' + json['type'])
-        return self.__run_actions(session_id, None, context, i - 1, verbose)
+        if current_request != self._sessions[session_id]:
+            return context
+        return self.__run_actions(session_id, current_request, None, context,
+                                  i - 1, verbose)
 
     def run_actions(self, session_id, message, context=None,
                     max_steps=DEFAULT_MAX_STEPS, verbose=None):
         if not self.actions:
             self.throw_must_have_actions()
-
         if context is None:
             context = {}
-        return self.__run_actions(session_id, message, context, max_steps, verbose)
+
+        # Figuring out whether we need to reset the last turn.
+        # Each new call increments an index for the session.
+        # We only care about the last call to run_actions.
+        # All the previous ones are discarded (preemptive exit).
+        current_request = self._sessions[session_id] + 1 if session_id in self._sessions else 1
+        self._sessions[session_id] = current_request
+
+        context = self.__run_actions(session_id, current_request, message,
+                                     context, max_steps, verbose)
+
+        # Cleaning up once the last call to run_actions finishes.
+        if current_request == self._sessions[session_id]:
+            del self._sessions[session_id]
+
+        return context
 
     def interactive(self, context=None, max_steps=DEFAULT_MAX_STEPS):
         """Runs interactive command line chat between user and bot. Runs