+++ CHANGES.md
@@ -1,3 +1,18 @@
+### Breaking changes
+
+- `message` now takes an optional `context` as second parameter
+
+## v4.2.0
+
+- added a `speech()` method to send audio files to the API (thanks @willywongi)
+
+## v4.1.0
+
+### Breaking changes
+
+- `converse` now takes `reset` as optional parameter.
+- `run_actions` now resets the last turn on new messages and errors.
+
 ## v4.0.0
 
 After a lot of internal dogfooding and bot building, we decided to change the API in a backwards-incompatible way. The changes are described below and aim to simplify user code and accommodate upcoming features.